Fixes on MinGW: - float_image was using a open-then-delete trick, which does not work,   just took it out for now.  We need to either fix it so that it does   work (which probably just means using the non-deprecated unlink on   windows, _unlink()), or delete them when we finally can. - Files were not being written as binary -- needs to be explicitly   specified.

+++ src/asf/tmpdir.c
@@ -71,12 +71,14 @@ int
 unlink_tmp_file(const char *filename)
 {
   char * full_name;
-  int ret;
+  int ret=0;
 
   full_name = full_tmp_name (filename);
-  ret = unlink (full_name);
-  if (ret < 0) 
-    asfPrintWarning("unlink_tmp_file '%s': %s\n", full_name, strerror(errno));
+  if (fileExists(full_name)) {
+    ret = unlink (full_name);
+    if (ret < 0) 
+      asfPrintWarning("unlink_tmp_file '%s': %s\n", full_name, strerror(errno));
+  }
   free(full_name);
 
   return ret;
